% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imputedata}
\alias{imputedata}
\title{Imputation results under different methods}
\format{
Data frame with 500 rows and 8 variables:
\itemize{
  \item MAR_T1: Average value of nonrespondents for continuous outcome in treatment group under MAR assumption.
  \item MAR_C1: Average value of nonrespondents for continuous outcome in control group under MAR assumption.
  \item MAR_T2: Number of success of nonrespondents for binary outcome in treatment group under MAR assumption.
  \item MAR_C2: Number of success of nonrespondents for binary outcome in control group under MAR assumption.
  \item MCAR_T1: Average value of nonrespondents for continuous outcome in treatment group under MCAR assumption.
  \item MCAR_C1: Average value of nonrespondents for continuous outcome in control group under MCAR assumption.
  \item MCAR_T2: Number of success of nonrespondents for binary outcome in treatment group under MCAR assumption.
  \item MCAR_C2:  Number of success of nonrespondents for binary outcome in control group under MCAR assumption.
}
}
\usage{
imputedata
}
\description{
Imputation results based on Missing At Random(MAR) and Missing Completely At Random(MCAR) assumption for treatment and control group.
}
\keyword{datasets}
