% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tippingdata}
\alias{tippingdata}
\title{An hypothetical dataset used to demonstrate functions.}
\format{
Data frame with 270 rows and 5 variables:
\itemize{
  \item continuous: continuous outcome
  \item binary: binary outcome
  \item educ: education time(years)
  \item female: 0=male, 1=female
  \item treat: 0=control group, 1=treatment group
}
}
\usage{
tippingdata
}
\description{
A hypothetical dataset with continuous and binary outcome.
}
\keyword{datasets}
